[RTOP-46] Separation of plc program and plc library
