+++ core/src/CMakeLists.txt
@@ -1,0 +1,31 @@
+cmake_minimum_required(VERSION 3.10)
+project(plc_application C)
+
+set(CMAKE_POSITION_INDEPENDENT_CODE ON)
+
+# Include directories
+include_directories(${CMAKE_SOURCE_DIR}/core/src/utils
+                    ${CMAKE_SOURCE_DIR}/core/generated/plc_lib
+                    ${CMAKE_SOURCE_DIR}/core/generated/plc_lib/lib)
+
+# Compiler options
+add_compile_options(-Wall -Werror -pedantic -Wextra -fstack-protector-strong
+    -D_FORTIFY_SOURCE=2 -O2 -Wformat
+    -Werror=format-security -fPIC -fPIE)
+
+# Step 3: Build the executable and link against the shared library
+add_executable(plc_main
+    ${CMAKE_SOURCE_DIR}/core/src/plc_app/plc_main.c
+    ${CMAKE_SOURCE_DIR}/core/src/plc_app/log.c
+    ${CMAKE_SOURCE_DIR}/core/src/plc_app/utils/utils.c
+    ${CMAKE_SOURCE_DIR}/core/src/plc_app/image_tables.c
+    ${CMAKE_SOURCE_DIR}/core/src/plc_app/watchdog.c
+)
+
+# Link against shared library
+target_link_libraries(plc_main plc dl pthread)
+
+# Ensure executable can find shared library at runtime
+set_target_properties(plc_main PROPERTIES
+    RUNTIME_OUTPUT_DIRECTORY ${CMAKE_BINARY_DIR}/bin
+)
