+++ core/src/plc_app/log.c
