+++ core/src/plc_app/utils/utils.c
